Move nipsa service to services package
