+++ h/services/__init__.py
@@ -6,4 +6,5 @@ from __future__ import unicode_literals
 
 
 def includeme(config):
+    config.register_service_factory('.nipsa.nipsa_factory', name='nipsa')
     config.register_service_factory('.settings.settings_factory', name='settings')
